refactor(article): update paths
